Ai-Agent-Design/ai-governance-agent/requirements.txt
# Azure SDKs
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# OpenAI
openai>=1.30.0

# Document processing
pymupdf>=1.23.0
python-docx>=1.1.0
tiktoken>=0.7.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0

# Utilities
requests>=2.31.0
rich>=13.7.0
